// components/bundle/menu.d.ts
export interface GlobalToolbarsWindow
{
	// Toolbars (Instances & Constructors)
	repositoryToolbar?: RepositoryToolbar;
	RepositoryToolbar?: typeof RepositoryToolbar;

	historyToolbar?: HistoryToolbar;
	HistoryToolbar?: typeof HistoryToolbar;

	settingsToolbar?: SettingsToolbar;
	SettingsToolbar?: typeof SettingsToolbar;

	scriptToolbar?: ScriptToolbar;
	ScriptToolbar?: typeof ScriptToolbar;

	appToolbar?: ApplicationToolbar;
	ApplicationToolbar?: typeof ApplicationToolbar;

	fileToolbar?: FileToolbar;
	FileToolbar?: typeof FileToolbar;

	// Managers & Widget Constructors
	settingsManager?: Settings;
	SettingsManager?: typeof Settings;

	fileManager?: FileManager;
	FileManager?: typeof FileManager;

	aceEditorWidget?: AceEditorWidget;
	AceEditorWidget?: typeof AceEditorWidget;

	spawnPoints: Record<string, string>;
	searchWorker?: Worker;

	searchService?: SearchService;
	SearchService?: typeof SearchService;

	githubService?: GithubService;
	GithubService?: typeof GithubService;

	layoutAdjuster?: LayoutAdjuster;
	LayoutAdjuster?: typeof LayoutAdjuster;
}