// components/bundle/make.d.ts
export interface BuildWindow
{
	SYS?: any;
	FILED?: any;
	FS: {
		virtual: Record<string, FileRecord | null | undefined>;
		pointers?: ([number, string, FileRecord, string, number, number])[];
	};
	path: typeof path;
	config: typeof config;
	dirs: typeof dirs;
	ENGINE_MEMORY_BASE?: number;
	UI_MEMORY_BASE?: number;
	CGAME_MEMORY_BASE?: number;
	GAME_MEMORY_BASE?: number;
	COMPILE_PLATFORM?: string;
	COMPILE_ARCH?: string;
	GAME_PLATFORM?: string;
	GAME_ARCH?: string;
	LONG_NAME?: string;
	SHORT_NAME?: string;
}